requirements.txt
bcrypt==3.1.1
boto==2.45.0
cffi==1.8.3
coverage==4.2
Django==1.7
django-braces==1.10.0
django-storages==1.5.1
Pillow==3.4.1
pycparser==2.17
six==1.10.0
